Check for defined 'nil' in var nil assignment inspection
